+++ src/vue/common/StatusMessage.vue
@@ -7,38 +7,54 @@
         `status-message--${messageType}`,
         { 'status-message--jump': isShownAgain }
       ]"
+      @mouseenter="stopProgressBar"
+      @mouseleave="startProgressBar"
     >
-      <div class="status-message__icon-wrp">
-        <i
-          class="mdi status-message__icon"
-          :class="[`status-message__icon--${messageType}`, messageIconClass]"
+      <div class="status-message__body">
+        <div class="status-message__icon-wrp">
+          <i
+            class="mdi status-message__icon"
+            :class="[`status-message__icon--${messageType}`, messageIconClass]"
+          />
+        </div>
+
+        <div class="status-message__payload">
+          <h4 class="status-message__title">
+            {{ 'status-message.title' | globalize({ context: messageType }) }}
+          </h4>
+
+          <p class="status-message__text">
+            <!-- eslint-disable-next-line max-len -->
+            {{ messageId | globalize({ context: messageType, ...messageArgs }) }}
+          </p>
+        </div>
+
+        <button
+          @click="isShown = false"
+          class="status-message__close-btn"
         />
       </div>
-
-      <div class="status-message__payload">
-        <h4 class="status-message__title">
-          {{ 'status-message.title' | globalize({ context: messageType }) }}
-        </h4>
-
-        <p class="status-message__text">
-          {{ messageId | globalize({ context: messageType, ...messageArgs }) }}
-        </p>
+      <div class="status-message__progress-bar">
+        <span
+          class="status-message__progress-bar-percentage"
+          :style="{ 'width': getInvertedProgressPercents + '%' }"
+          :class="`status-message__progress-bar-percentage--${messageType}`"
+        />
       </div>
-
-      <button
-        @click="isShown = false"
-        class="status-message__close-btn"
-      />
     </div>
   </transition>
 </template>
 
 <script>
 import { Bus } from '@/js/helpers/event-bus'
+import { MathUtil } from '@/js/utils'
 
 const DEFAULT_MESSAGE_TRANSLATION_ID = 'status-message.default-message'
 const CLOSE_TIMEOUT_MS = 10000
 const SHOWN_AGAIN_PRESENCE_TIMEOUT_MS = 1000
+const ONE_HUNDRED_PERCENTS = 100
+const MAX_ANIMATION_PROGRESS = 1
+const MIN_ANIMATION_PROGRESS = 0
 const MESSAGE_TYPES = Object.freeze({
   warning: 'warning',
   success: 'success',
@@ -56,8 +72,12 @@ export default {
     messageArgs: {},
     isShown: false,
     isShownAgain: false,
-    isShownTimeoutId: -1,
     isShownAgainTimeoutId: -1,
+    animationFrame: -1,
+    progressBar: {
+      progress: 0,
+      paused: false,
+    },
   }),
 
   computed: {
@@ -75,6 +95,19 @@ export default {
           return ''
       }
     },
+
+    getInvertedProgressPercents () {
+      const animationProgressPercents = MathUtil.multiply(
+        this.progressBar.progress,
+        ONE_HUNDRED_PERCENTS
+      )
+
+      const animationProgressPercentsInverted = MathUtil.subtract(
+        ONE_HUNDRED_PERCENTS,
+        animationProgressPercents
+      )
+      return animationProgressPercentsInverted
+    },
   },
 
   created () {
@@ -86,6 +119,10 @@ export default {
       this.show(MESSAGE_TYPES.error, payload))
     Bus.on(Bus.eventList.info, payload =>
       this.show(MESSAGE_TYPES.info, payload))
+  },
+
+  destroyed () {
+    cancelAnimationFrame(this.animationFrame)
   },
 
   methods: {
@@ -101,19 +138,51 @@ export default {
         this.messageId = DEFAULT_MESSAGE_TRANSLATION_ID
       }
 
+      cancelAnimationFrame(this.animationFrame)
+
       if (this.isShown) {
         this.doJump()
+        this.progressBar.progress = MIN_ANIMATION_PROGRESS
       } else {
         this.isShown = true
       }
 
-      if (this.isShownTimeoutId >= 0) {
-        window.clearTimeout(this.isShownTimeoutId)
-      }
+      this.startAnimationTimeout()
+    },
 
-      this.isShownTimeoutId = window.setTimeout(_ => {
-        this.isShown = false
-      }, CLOSE_TIMEOUT_MS)
+    startAnimationTimeout (currentAnimationTime = 0) {
+      const animationStartTime = performance.now()
+      const calculateAnimationProgress = () => {
+        this.animationFrame = requestAnimationFrame((timestamp) => {
+          if (this.progressBar.paused) {
+            cancelAnimationFrame(this.animationFrame)
+            return
+          }
+
+          const animationTime = MathUtil.subtract(
+            currentAnimationTime,
+            animationStartTime
+          )
+          const animationRuntime = MathUtil.add(
+            timestamp,
+            animationTime
+          )
+          const animationProgress = MathUtil.divide(
+            animationRuntime,
+            CLOSE_TIMEOUT_MS
+          )
+
+          if (animationRuntime < CLOSE_TIMEOUT_MS) {
+            this.progressBar.progress = animationProgress
+            calculateAnimationProgress()
+          } else {
+            this.progressBar.progress = MAX_ANIMATION_PROGRESS
+            cancelAnimationFrame(this.animationFrame)
+            this.isShown = false
+          }
+        })
+      }
+      calculateAnimationProgress()
     },
 
     doJump () {
@@ -130,6 +199,20 @@ export default {
 
     isObject (value) {
       return typeof value === 'object' && !Array.isArray(value) && !null
+    },
+
+    stopProgressBar () {
+      this.progressBar.paused = true
+    },
+
+    startProgressBar () {
+      this.progressBar.paused = false
+      const currentAnimationTime = MathUtil.multiply(
+        CLOSE_TIMEOUT_MS,
+        this.progressBar.progress
+      )
+
+      this.startAnimationTimeout(currentAnimationTime)
     },
   },
 }
@@ -151,6 +234,7 @@ $icon-padding: 2.4rem;
   max-width: 42rem;
   min-width: 32rem;
   display: flex;
+  flex-direction: column;
 
   &--warning {
     background-color: $col-status-msg-warning;
@@ -178,6 +262,43 @@ $icon-padding: 2.4rem;
   @include respond-to-custom($status-message-reposition-bp) {
     right: $content-side-paddings-sm;
     left: $content-side-paddings-sm;
+  }
+}
+
+.status-message__body {
+  display: flex;
+}
+
+.status-message__progress-bar {
+  position: absolute;
+  width: 100%;
+  height: 0.5rem;
+  bottom: 0;
+}
+
+.status-message__progress-bar-percentage {
+  position: absolute;
+  top: 0;
+  left: 0;
+  height: 0.5rem;
+  width: 100%;
+  background-color: $col-status-msg-success;
+  max-width: 100%;
+
+  &--success {
+    background-color: $col-status-msg-success;
+  }
+
+  &--error {
+    background-color: $col-status-msg-text;
+  }
+
+  &--info {
+    background-color: $col-status-msg-info;
+  }
+
+  &--warning {
+    background-color: $col-status-msg-text;
   }
 }