+++ CHANGELOG.md
@@ -28,6 +28,8 @@ for further information about branching and tagging conventions.
   - `OFFLINE_ISSUANCE_MAC_LINK`
   - `OFFLINE_ISSUANCE_SOURCE_LINK`
 - Empty message placeholder for My assets page
+- `ErrorTracker.setLoggedInUser`to `SET_ACCOUNT`, `CLEAR_STATE` and `SET_WALLET`
+  mutations
 
 #### Changed
 - If some of the config keys are empty, the related images and links will not
@@ -41,6 +43,7 @@ for further information about branching and tagging conventions.
 
 #### Removed
 - `RECOVERY_MODE` config key due to unused anymore
+- `ErrorTracker.setLoggedInUser` from App.vue and LoginForm.vue
 
 ## [1.10.0-rc.3] - 2019-07-22
 #### Added
@@ -52,6 +55,7 @@ for further information about branching and tagging conventions.
 #### Fixed
 - Issue when a user could not create an atomic swap if an amount is less than
   an available balance
+- Overflow clipboard-field on tfa form
 
 ## [1.10.0-rc.2] - 2019-07-18
 #### Added