+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
-## [Unreleased]
+## [1.10.0-rc.4] - 2019-08-02
 #### Fixed
 - A bug when you cannot return from "Downloads" if the tab was opened in new tab
 - A bug when page reload twice
@@ -956,7 +956,8 @@ for further information about branching and tagging conventions.
 
 ## [1.3.0] - 2019-03-01
 
-[Unreleased]: https://github.com/tokend/web-client/compare/1.10.0-rc.3...HEAD
+[Unreleased]: https://github.com/tokend/web-client/compare/1.10.0-rc.4...HEAD
+[1.10.0-rc.4]: https://github.com/tokend/web-client/compare/1.10.0-rc.3...1.10.0-rc.4
 [1.10.0-rc.3]: https://github.com/tokend/web-client/compare/1.10.0-rc.2...1.10.0-rc.3
 [1.10.0-rc.2]: https://github.com/tokend/web-client/compare/1.10.0-rc.1...1.10.0-rc.2
 [1.10.0-rc.1]: https://github.com/tokend/web-client/compare/1.10.0-rc.0...1.10.0-rc.1